Finished refractoring

+++ src/app/community/community.component.ts
@@ -2,7 +2,6 @@ import { Component, OnInit, AfterViewInit, ViewChild } from '@angular/core';
 import { Location } from '@angular/common';
 import { ActivatedRoute, ParamMap } from '@angular/router';
 import { CouchService } from '../shared/couchdb.service';
-import { DialogsDeleteComponent } from '../shared/dialogs/dialogs-delete.component';
 import { DialogsEditComponent } from '../shared/dialogs/dialogs-edit.component';
 import { MatTableDataSource, MatPaginator, MatDialog } from '@angular/material';
 import { switchMap } from 'rxjs/operators';
@@ -62,96 +61,36 @@ export class CommunityComponent implements OnInit, AfterViewInit {
       }, (error) => this.message = 'There was a problem getting Communities');
   }
 
-  deleteClick(community) {
-    this.deleteDialog = this.dialog.open(DialogsDeleteComponent, {
-      data: {
-        okClick: this.deleteCommunity(community),
-        type: 'community',
-        displayName: community.name
-      }
-    });
-  }
-
-  acceptClick(community) {
+  updateClick(community, change) {
     this.editDialog = this.dialog.open(DialogsEditComponent, {
       data: {
-        okClick: this.acceptCommunity(community),
+        okClick: this.updateCommunity(community, change),
+        changeType: change,
         type: 'community',
         displayName: community.name
       }
     });
   }
 
-  rejectClick(community) {
-    this.editDialog = this.dialog.open(DialogsEditComponent, {
-      data: {
-        okClick: this.rejectCommunity(community),
-        type: 'community',
-        displayName: community.name
-      }
-    });
-  }
-
-  unlinkClick(community) {
-    this.editDialog = this.dialog.open(DialogsEditComponent, {
-      data: {
-        okClick: this.unlinkCommunity(community),
-        type: 'community',
-        displayName: community.name
-      }
-    });
-  }
-
-  deleteCommunity(community) {
+  updateCommunity(community, change) {
     // Return a function with community on its scope to pass to delete dialog
     return () => {
     // With object destructuring colon means different variable name assigned, i.e. 'id' rather than '_id'
       const { _id: id, _rev: rev } = community;
-      this.couchService.delete('communityregistrationrequests/' + id + '?rev=' + rev)
+      if (change !== 'delete') {
+        community['registrationRequest'] = change;
+        this.couchService.put('communityregistrationrequests/' + id + '?rev=' + rev, community)
+        .subscribe((data) => {
+          this.editDialog.close();
+        }, (error) => this.editDialog.componentInstance.message = 'There was a problem accepting this community');
+      } else {
+        this.couchService.delete('communityregistrationrequests/' + id + '?rev=' + rev)
         .subscribe((data) => {
           // It's safer to remove the item from the array based on its id than to splice based on the index
           this.communities.data = this.communities.data.filter((comm: any) => data.id !== comm._id);
           this.deleteDialog.close();
         }, (error) => this.deleteDialog.componentInstance.message = 'There was a problem deleting this community');
-    };
-  }
-
-  acceptCommunity(community) {
-    // Return a function with community on its scope to pass to delete dialog
-    return () => {
-    // With object destructuring colon means different variable name assigned, i.e. 'id' rather than '_id'
-      const { _id: id, _rev: rev } = community;
-      community['registrationRequest'] = 'accepted';
-      this.couchService.put('communityregistrationrequests/' + id + '?rev=' + rev, community)
-        .subscribe((data) => {
-          this.editDialog.close();
-        }, (error) => this.editDialog.componentInstance.message = 'There was a problem accepting this community');
-    };
-  }
-
-  rejectCommunity(community) {
-    // Return a function with community on its scope to pass to delete dialog
-    return () => {
-    // With object destructuring colon means different variable name assigned, i.e. 'id' rather than '_id'
-      const { _id: id, _rev: rev } = community;
-      community['registrationRequest'] = 'rejected';
-      this.couchService.put('communityregistrationrequests/' + id + '?rev=' + rev, community)
-        .subscribe((data) => {
-          this.editDialog.close();
-        }, (error) => this.editDialog.componentInstance.message = 'There was a problem rejecting this community');
-    };
-  }
-
-  unlinkCommunity(community) {
-    // Return a function with community on its scope to pass to delete dialog
-    return () => {
-    // With object destructuring colon means different variable name assigned, i.e. 'id' rather than '_id'
-      const { _id: id, _rev: rev } = community;
-      community['registrationRequest'] = 'unlinked';
-      this.couchService.put('communityregistrationrequests/' + id + '?rev=' + rev, community)
-        .subscribe((data) => {
-          this.editDialog.close();
-        }, (error) => this.editDialog.componentInstance.message = 'There was a problem unlinking this community');
+      }
     };
   }
